Incorporate PR comments
