+++ README.md
@@ -215,8 +215,6 @@ match(pet, Pet(_, 7), lambda name: name)                        # => 'rover'
 match(pet, Pet(_, _), lambda name, age: (name, age))            # => ('rover', 7)
 ```
 
-It is possible to match on different dataclasses withing the same invocation.
-
 ## Install
 
 Currently it works only in Python >= 3.6 [Because dict matching can work only in the latest Pythons](https://mail.python.org/pipermail/python-dev/2017-December/151283.html).
