+++ pampy/pampy.py
@@ -63,7 +63,7 @@ def match_value(pattern, value) -> Tuple[bool, List]:
         return True, [value]
     elif pattern is HEAD or pattern is TAIL:
         raise MatchError("HEAD or TAIL should only be used inside an Iterable (list or tuple).")
-    elif is_dataclass(pattern) and (pattern.__class__ == value.__class__):
+    elif is_dataclass(pattern) and pattern.__class__ == value.__class__:
         return match_dict(pattern.__dict__, value.__dict__)
     return False, []
